==> codes/1.concat_preprocessing_data(withnewlabel).ipynb <==
from numpy import array
from numpy import asarray
from numpy import zeros
import pandas as pd
import numpy as np

import re
import pickle 
#import mglearn
import time
# import gensim
import os
import collections
# import smart_open
import random
import string
import nltk

from nltk import Text
from nltk.tokenize import TweetTokenizer # doesn't split at apostrophes
from nltk.tokenize import regexp_tokenize
from nltk.tokenize import word_tokenize  
from nltk.tokenize import sent_tokenize 
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from nltk.stem import PorterStemmer

from sklearn import decomposition
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.preprocessing import LabelEncoder
from sklearn.metrics import classification_report, confusion_matrix, accuracy_score, f1_score

import preprocessor as p

import keras.backend as K
import keras.layers as layers
import numpy as np
import pandas as pd
import tensorflow as tf
import tensorflow_hub as hub
from keras.callbacks import ModelCheckpoint
from keras.engine import Layer
from keras.layers import CuDNNLSTM, Bidirectional, GlobalMaxPooling1D, GlobalAveragePooling1D
from keras.layers import Embedding, Dense, Flatten, Input
from keras.layers import SpatialDropout1D, add, concatenate
from keras.layers.convolutional import Conv1D
from keras.layers.pooling import MaxPool1D
from keras.models import Model
from keras.preprocessing import text, sequence


from keras.preprocessing.text import Tokenizer
from keras.preprocessing.sequence import pad_sequences
from keras.utils import to_categorical

from sklearn.preprocessing import LabelEncoder
from sklearn.metrics import accuracy_score
from sklearn.metrics import classification_report
from sklearn.model_selection import train_test_split


from hltc_preprocess import tweets
import tweepy
import pandas as pd

import glob

# 0. 같은 감정 데이터들 하나로 묶고 저장

anticipation은 할 필요 없음 

def data_load(file_path, emo):
    files = glob.glob(file_path+'tweet_data/{0}/*.csv'.format(emo))
    
    all_emo_data = []
    for file in files:     
        data = pd.read_csv(file, names=['text'], header=0)
        all_emo_data.append(data)
        
    concated_data = pd.concat(all_emo_data)
    concated_data.to_csv(file_path+"tweet_data/{0}/{0}_concat_data.csv".format(emo))
    print(emo,"concat data is saved")
    return concated_data

if __name__ == "__main__":
    emo = 'anticipation'
    file_path = 'C:/Users/Administrator/Anaconda3/envs/temp/EmotionEmbedding/emotion8_finalcode/data/'
    concated_data = data_load(file_path, emo)

# 1. each emo concat data labeling // all emo concat

def concat_data_labeling(path, emo, emo_label_dic):
    file_emo = glob.glob(path)
    data = pd.read_csv(file_emo[0], header=0)
    data = data.reset_index(drop= True)
    data = data.dropna()
    label_num = emo_label_dic[emo]
    data['label'] = label_num
#     display(data.tail(10))
    
    return data

# function for cleaning the plots of the movies
def clean_text(text):
    text = text.lower()
    ## 특수기호제거 
    text = re.sub('[^A-Za-z ]', '',text)
    text = re.sub(r'[0-9]+', '', text)
    text = re.sub(r"what's", "what is ", text)
    text = re.sub(r"\'s", " ", text)
    text = re.sub(r"\'ve", " have ", text)
    text = re.sub(r"can't", "can not ", text)
    text = re.sub(r"n't", " not ", text)
    text = re.sub(r"i'm", "i am ", text)
    text = re.sub(r"\'re", " are ", text)
    text = re.sub(r"\'d", " would ", text)
    text = re.sub(r"\'ll", " will ", text)
    text = re.sub(r"\'scuse", " excuse ", text)
    #text = re.sub('\W', ' ', text)
    #text = re.sub('\s+', ' ', text)
    text = text.strip(' ')
    return text

def tweet_preprocessing(data):
    emo_data_clean = data.copy()
    emo_data_clean['clean_text'] = emo_data_clean['text'].apply(tweets.clean_tweet)
    emo_data_clean['clean_text'] = emo_data_clean['clean_text'].apply(clean_text)
    emo_data_clean = emo_data_clean[['clean_text','label']]
    return emo_data_clean

if __name__ == "__main__":
    emo_list = ['anger','anticipation','disgust','fear','joy','trust','sadness','surprise']
    label_list = list(range(len(emo_list)))
    emo_label_dic = dict(zip(emo_list, label_list))
    
    all_emo_clean_data = []
    file_path = 'C:/Users/Administrator/Anaconda3/envs/temp/EmotionEmbedding/emotion8_finalcode/data/'
    
    for emo in emo_list:
        path = file_path+'tweet_data/{0}/{0}_concat_data.csv'.format(emo)
        data = concat_data_labeling(path, emo, emo_label_dic)
        
        emo_data_clean = tweet_preprocessing(data)
        print(emo, "data shape : ", emo_data_clean.shape)
        
        emo_data_clean.to_csv(file_path+"tweet_data/{0}/{0}_emo_clean_data3.csv".format(emo))
        all_emo_clean_data.append(emo_data_clean)
        
    concated_emo_clean_data = pd.concat(all_emo_clean_data)
    concated_emo_clean_data.to_csv(file_path+"tweet_data/all_emo_clean_data/all_emo_clean_data3.csv")

result = pd.read_csv(file_path+"tweet_data/all_emo_clean_data/all_emo_clean_data3.csv")
result

# Tweet data label check 

import pandas as pd

# final label check 
file_path = 'C:/Users/Administrator/Anaconda3/envs/temp/EmotionEmbedding/emotion8_finalcode/data/'
final_data = pd.read_csv(file_path+"tweet_data/all_emo_clean_data/all_emo_clean_data3.csv")

final_data['label'].value_counts()

for i in range(8):
    display(final_data[final_data['label']==i].head())
