==> codes/3.getSimilarity_ROCandHashwords.ipynb <==
import pandas as pd 
import numpy as np 
import pickle 

import numpy as np
import pandas as pd
import re
# from lxml import etree
import nltk
from nltk.tokenize import word_tokenize, sent_tokenize
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.metrics.pairwise import linear_kernel
from sklearn.metrics.pairwise import cosine_similarity
import numpy as np
import os
import pickle

import konlpy
import jpype
from konlpy.tag import Okt
from konlpy.tag import Komoran
from konlpy.tag import Twitter
from gensim.models.word2vec import Word2Vec
from gensim.models.doc2vec import *
from gensim.models import KeyedVectors
from gensim.models import FastText

finalresult = file_path+"finalresult/result%s/result_%s.csv"%(num,name)
finalresult

name = 'pe8_3'
num = name[4:]

file_path = 'C:/Users/Administrator/Anaconda3/envs/temp/EmotionEmbedding/emotion8_finalcode/'
em_layername = file_path+"embeddinglayer/layer_%s.pickle"%name # keras embedding layer name 
glove_filename = file_path+"glovematrix/glovemat_%s.pickle"%name  # save glove embedding matrix 
word_indexname = file_path+"wordindex/wordindex_%s.pickle"%name  # save glove embedding matrix 
roc_datapath = file_path+"data/rocdata/rocstories_sentword_lemma_31m_final.csv"


finalresult = file_path+"finalresult/result%s/result_%s.csv"%(num,name)  ## roc emotion mapping result path 
tweetlabel_datapath = file_path+'data/tweet_labelwords'

def all_files_load():
    with open(em_layername, 'rb') as rotten_file:
        embeddinglayer = pickle.load(rotten_file)
    
    with open(glove_filename, 'rb') as rotten_file:
        glovematrix = pickle.load(rotten_file)
    
    with open(word_indexname, 'rb') as rotten_file:
        wordindex = pickle.load(rotten_file)
        
    roc_data = pd.read_csv(roc_datapath,names = ['storyid','sequence','text','sentword'])
    
    return embeddinglayer, glovematrix, wordindex, roc_data

def get_tweet_labelwords():
    with open(tweetlabel_datapath, 'rb') as rotten_file:
        tweet_labeldict = pickle.load(rotten_file)
        
    return tweet_labeldict

def get_tweetvectormatrix(tweet_labelwordlist, finaldic):
    # get tweet data vector 
    final_tweet_labellist = []
    wordvec = np.zeros(300,)
    for word in tweet_labelwordlist:
        try :
            wordvec = np.vstack([wordvec,finaldic[word]])
            final_tweet_labellist.append(word)
        except KeyError:
            print(word,'is not in embedding matrix')
    
    wordvec = wordvec[1:]
    tweet_vec_mat = wordvec
    return tweet_vec_mat, final_tweet_labellist

if __name__ == "__main__":
    embeddinglayer, glovematrix, wordindex, roc_data = all_files_load()
    tweet_labeldict = get_tweet_labelwords()
    tweet_labelwordlist =  list(set(sum(list(tweet_labeldict.values()), [])))
    
    numlist = list(range(8))
    newdic = dict(zip(numlist,tweet_labeldict.values()))
    tweet_emo_final_dic = {}
    for key,emolist in newdic.items():
        for emo in emolist:
            tweet_emo_final_dic[emo] = key

    
    # dictionary = word : wordvector
    global finaldic
    finaldic = {w:embeddinglayer[idx] for w, idx in wordindex.items()}
    
    global tweet_vec_mat
    tweet_vec_mat, final_tweet_labellist = get_tweetvectormatrix(tweet_labelwordlist,finaldic)
    # roc_data['sentword_vector'] = roc_data['sentword'].apply(get_rocstory_vector)
    
#     tweet_vec_mat = tweet_vec_mat
    
    

embeddinglayer.shape

# emotion labeling using cosine similarity

def get_rocstory_vector(inputword):
    try :
        inputwordvec = finaldic[inputword]
        cosine_similarity = linear_kernel([inputwordvec], tweet_vec_mat)
        
        simvalue = cosine_similarity[0]
        simvalue_windex = list(enumerate(simvalue))
        
        #print(simvalue_windex)
        
        sim_scores = sorted(simvalue_windex, key=lambda x: x[1], reverse=True)
        sim_scores_top1 = sim_scores[0]
        emotion = final_tweet_labellist[sim_scores_top1[0]]
        emotionlabel = tweet_emo_final_dic[emotion]
        
        sim_scores_top_5 = sim_scores[0:5]
        sim_scores_top_5_index = [i[0] for i in sim_scores_top_5]
        emotionlist = [final_tweet_labellist[i] for i in sim_scores_top_5_index]
        
        return emotionlist, emotionlabel

    except KeyError:
        return 0, None

if __name__ == "__main__": 
    
    roc_data['sentword_vector'],roc_data['emotionlabel'] = zip(*roc_data['sentword'].apply(get_rocstory_vector))
    roc_data_dropna = roc_data.dropna(subset = ['emotionlabel'])
    roc_data_dropna.to_csv(finalresult)

# result

roc_data_dropna

roc_data_dropna['emotionlabel'].value_counts()

### check
roc_data_dropna[roc_data_dropna['emotionlabel']==5]

# (ignore) Analysispart

numlist = list(range(8))
newdic = dict(zip(numlist,tweet_labeldict.values()))
tweet_emo_final_dic = {}
for key,emolist in newdic.items():
    for emo in emolist:
        tweet_emo_final_dic[emo] = key

newdic
