==> codes/4.ResultAnalysis(1)_pe8_3.ipynb <==
import pandas as pd 
import numpy as np 
import collections

def report_load(finalresult):
    final_df = pd.read_csv(finalresult, index_col = 0) 
    final_df = final_df.reset_index(drop = True)
    return final_df

if __name__ == '__main__':
    ## path
    file_path = 'C:/Users/Administrator/Anaconda3/envs/temp/EmotionEmbedding/emotion8_finalcode/finalresult/result3/'
    name = 'pe8_3'
    finalresult = file_path+"result_%s.csv"%name # final emotion label result csv
    
    
    ## load data
    final_df = report_load(finalresult)
    print(final_df['emotionlabel'].value_counts())

final_df

# 각 감정 개수 확인 및 문장 뽑기 

def check_emotionlabel(emo_list):
    result_dic = final_df['emotionlabel'].value_counts()
    result_df = pd.DataFrame.from_dict(result_dic)
    result_df = result_df.sort_index()
    
    new = pd.DataFrame()
    new['emotion'] = emo_list
    new['count'] = result_df['emotionlabel']
    new['per'] = [round((i/sum(new['count']))*100,2) for i in new['count']]
    return new

def get_emosentence(emo_list, file_path):
    for i in range(8):
        one_emo_result = final_df[final_df['emotionlabel'] == i]
        one_emo_result = one_emo_result.reset_index(drop=True)
#         one_emo_result = 
        display(one_emo_result[:10])
        one_emo_result.to_csv(file_path+'{0}_emo_result.csv'.format(emo_list[i]))
        print(i, emo_list[i],'isdone')

if __name__ == "__main__": 
    
    emo_list = ['anger','anticipation','disgust','fear','joy', 'trust','sadness','surprise']
    new = check_emotionlabel(emo_list)
    display(new)
    result_table = check_emotionlabel(emo_list)
    get_emosentence(emo_list,file_path)

# 각 감정별 키워드 랭킹 뽑기 

def get_sentiword_ranking(word_list):
    emo_word_dict = dict(collections.Counter(word_list))
    emo_rank = pd.DataFrame()
    emo_rank['word'] = emo_word_dict.keys()
    emo_rank['counts'] = emo_word_dict.values()
    emo_rank = emo_rank.sort_values(by = 'counts',ascending  = False)
    emo_rank = emo_rank.reset_index(drop = True)
    
    return emo_rank

if __name__ == "__main__":
    emo_list = ['anger','anticipation','disgust','fear','joy', 'trust', 'sadness','surprise']
    for i in range(8):
        print(i, emo_list[i])
        word_list = list(final_df['sentword'][final_df['emotionlabel'] == i].values)
        emo_rank = get_sentiword_ranking(word_list)
        display(emo_rank[:20])
