==> codes/4.ResultAnalysis(2)_pe8_3.ipynb <==
import pandas as pd
import numpy as np
import glob

# 1. 각 data set 에서 랜덤하게 10 문장 뽑기 
# 2. sent word 결정적인거 기반으로 5문장 임의로 뽑기  ==> 이건 수작업
# 3. 겹치는거 있는 지 확인하고 
# 4. 전체 랜덤하게 섞기 

## 1. 각 data set 에서 랜덤하게 10 문장 뽑기 

def data_load(files):
    all_data = []
    for file in files:
        raw_data = pd.read_csv(file,header=0)
        data = raw_data.iloc[:,2:]
        all_data.append(data)
        
    return all_data

def random_sampling(file_path, emo_list, all_data):
    all_sample_list = []
    for idx, data in enumerate(all_data):
        
        # story title 은 샘플에서 제외 
        data = data[data['sequence']!= 'storytitle']
        sample = data.sample(n=10)
        
        # sample csv로 저장하기
        sample.to_csv(file_path+'sample/{0}_emo_sample.csv'.format(emo_list[idx]))
        print(idx, emo_list[idx],'isdone')
        all_sample_list.append(sample)
                
    return all_sample_list 

def concat_and_suffle_data(all_sample_list):
    all_sample_df  = pd.concat(all_sample_list)
    all_sample_df = all_sample_df.sample(frac = 1).reset_index(drop = True)
    
    return all_sample_df

if __name__ == "__main__":
    emo_list = ['anger','anticipation','disgust','fear','joy', 'trust','sadness','surprise']
    
    file_path = 'C:/Users/Administrator/Anaconda3/envs/temp/EmotionEmbedding/emotion8_finalcode/finalresult/result3/'
    file_name = '*_emo_result.csv'
    files = glob.glob(file_path+file_name)
        
    all_data = data_load(files) # load daa
    all_sample_list = random_sampling(file_path, emo_list, all_data)

    all_sample_df = concat_and_suffle_data(all_sample_list)
    all_sample_df.to_csv(file_path+"sample/all_sample_df_suffle.csv")

all_sample_df

## 2. sent word 결정적인거 기반으로 5문장 임의로 뽑은것 + 아까 셔플된 샘플 합쳐서 다시 섞기 

file_path = 'C:/Users/Administrator/Anaconda3/envs/temp/EmotionEmbedding/emotion8_finalcode/finalresult/result3/'
file_name = '*_emo_result.csv'
selec_file_name = 'sample/my_select_emo.csv'

def data_load_selected(file_path,selec_file_name):
    my_data = pd.read_csv(file_path+selec_file_name, index_col = 0)
    my_data = my_data.iloc[:,1:]
    return my_data

if __name__ == "__main__":
    my_data = data_load_selected(file_path,selec_file_name)
    total_sample_list = [my_data, all_sample_df]
    
    total_sample_df = concat_and_suffle_data(total_sample_list)
    total_sample_df.to_csv(file_path+"sample/total_sample_df_suffle.csv")
    

total_sample_df

total_sample_df

file_path = 'C:/Users/Administrator/Anaconda3/envs/temp/EmotionEmbedding/emotion8_finalcode/finalresult/result3/sample/'
total_sample_df = pd.read_csv(file_path+"sample/total_sample_df_suffle.csv",index_col = 0 )
total_sample_df['num_col'] =  [str(i) for i in range(1, len(total_sample_df)+1)]
total_sample_df['text+num'] = total_sample_df['num_col']+"."+" "+total_sample_df['text']
del total_sample_df['num_col'] 
total_sample_df_survey = total_sample_df.copy()
total_sample_df_survey.to_csv(file_path+"total_sample_df_suffle_survey.csv")

total_sample_df_survey
