==> codes/5.userrateanalysis-acc.ipynb <==
import pandas as pd
import numpy as np
import glob
import collections
from collections import Counter

def data_load():
    user_data = pd.read_csv("roc_humanrater.csv")
    question_data = pd.read_csv("total_sample_df_suffle_survey_edited.csv")
    
    return user_data, question_data

if __name__ == "__main__":
    user_data, question_data = data_load()

user_data_emo = user_data[1::2]
user_data_pos = user_data[0::2]

map_dic = {'Anger':0, 'Anticipation':1,'Disgust':2, 'Fear':3, 'Joy':4, 'Trust':5,'Sadness':6,'Surprise':7}
func = lambda x: map_dic.get(x, x)

user_data_emo_label  = user_data_emo.copy()
user_data_emo_label['user1'] = user_data_emo.user1.map(func)
user_data_emo_label['user2'] = user_data_emo.user2.map(func)
user_data_emo_label['user3'] = user_data_emo.user3.map(func)
user_data_emo_label['user4'] = user_data_emo.user4.map(func)

user_data_emo_label = user_data_emo_label.reset_index(drop = True)
user_data_emo_label['label'] = question_data['emotionlabel']
sort_by_label = user_data_emo_label.sort_values(by = ['label'])

total_human_rate_list = []
for i in range(8):
    human_rate_list = list(sort_by_label[sort_by_label['label']==i].iloc[:,1:-1].values.flatten())
    total_human_rate_list.append(human_rate_list)

sort_by_label[sort_by_label['label']==0]

## check content 
emo_list = ['anger','anticipation','disgust','fear','joy', 'trust','sadness','surprise']

for i in range(8):
    print(emo_list[i])
    display(sort_by_label[sort_by_label['label']==i].head(10))

total_correct_ratio = {}
emo_list = ['anger','anticipation','disgust','fear','joy', 'trust','sadness','surprise']

for i in range(8):
    correct_label = total_human_rate_list[i].count(i) 
    correct_ratio = round(correct_label/len(total_human_rate_list[i]),3)
    total_correct_ratio[emo_list[i]] = correct_ratio


total_correct_ratio

for i in range(8):
    tt = dict(Counter(total_human_rate_list[i]))
    print(i, sorted(tt.items()))

user_data_pos.head()

map_dic = {'negative':0, 'positive':1,'neutral':2}
func2 = lambda x: map_dic.get(x, x)

user_data_pos_label  = user_data_pos.copy()
user_data_pos_label['user1'] = user_data_pos.user1.map(func2)
user_data_pos_label['user2'] = user_data_pos.user2.map(func2)
user_data_pos_label['user3'] = user_data_pos.user3.map(func2)
user_data_pos_label['user4'] = user_data_pos.user4.map(func2)

user_data_pos_label.head()

user_data_pos_label = user_data_pos_label.reset_index(drop = True)
user_data_pos_label['label'] = question_data['emotionlabel']
sort_by_label_pos = user_data_pos_label.sort_values(by = ['label'])

sort_by_label_pos.head()

total_correct_ratio = {}
# emo_list = ['anger','anticipation','disgust','fear','joy', 'trust','sadness','surprise']

total_pos_count = []
for i in range(8):
    human_rate_list_pos = list(sort_by_label_pos[sort_by_label_pos['label']==i].iloc[:,1:-1].values.flatten())
    one_human_rate_series_pos = pd.Series(human_rate_list_pos)
    one_pos_count= one_human_rate_series_pos[one_human_rate_series_pos.isin(flags)].value_counts().reindex(flags, fill_value=0)
    if i == 0:
        df = one_pos_count.to_frame().T
    else: 
        onedf = one_pos_count.to_frame().T
        df = pd.merge(df,onedf,how='outer')

df.to_csv("pos_result.csv")

count_dic = Counter(total_human_rate_list_pos[0])
pd.DataFrame.from_dict(count_dic, orient  = 'index')


total_human_rate_list_pos
